Update MQ2AdvLootType.cpp
Added Status the to TLO AdvLoot.SList TLO. If there is a an item in your advloot window and you do a ${AdvLoot.SList[1].Status} it will return a string of the status of that item in the window. Which are WAITING, ASKING, ROLLING, STOPPED, CLICKROLL, FREEGRAB, and UNKNOWN just for future proofing in case other statuses happen in the future. If you use the Personal loot list it will return INVALID. If you use an invalid slot ID it will return null. Thanks to PhantomRogue for letting me bounce idea off of him.